Refactor Visitors to one component per file
