+++ examples/demo/src/segments/LinkToRelatedCustomers.js
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 import { translate } from 'react-admin';
 import { stringify } from 'query-string';
 
-import { VisitorIcon } from '../visitors';
+import visitors from '../visitors';
 
 const styles = {
     icon: { paddingRight: '0.5em' },
@@ -31,7 +31,7 @@ const LinkToRelatedCustomers = ({ classes, segment, translate }) => (
         }}
         className={classes.link}
     >
-        <VisitorIcon className={classes.icon} />
+        <visitors.icon className={classes.icon} />
         {translate('resources.segments.fields.customers')}
     </Button>
 );
